Copr building: default to dry-run; allow project override (#34)
For safety, default to dry-run mode where only the commands that would
run are echoed.  If you want to actually run them use
"make DRYRUN=false ...".

Allow overriding of the COPR_PROJECT either in copr-local.mk or on the
make commandline with "make COPR_PROJECT=<project_name> ...".

Allow forcing the building of an SRPM (i.e. vs. just the spec) by setting
UNPUBLISHED=true (i.e. "make UNPUBLISHED=true ...".  This is needed when
building a module type that usually gets it's sources from a registry, like
npm or pypi, etc. and the version you want to build is not yet published,
such as would be the case when testing a module build prior to landing.

Signed-off-by: Brian J. Murrell <brian.murrell@intel.com>