Allow a package to opt out of rpmlint
Because we could be building other people's specfiles and
they may not be requiring rpmlint to pass.

Also, some specfiles are impossible to pass rpmlint with because
rpmlint does not support setting macros with.

But we still do always rpmlint, just for information.  We just
don't fail if the package has opted out.

Signed-off-by: Brian J. Murrell <brian.murrell@intel.com>