Get version and release from specfile
If not using an RPM from a repo, assume there is a specfile that
is canonical and that we can get the desired version and release
from it.

A PACKAGE_VERSION and PACKAGE_RELEASE in the Makefile still over-
ride the specfile (but don't update the specfile) Version and
Release however.  This is not something you will likely want to
do.

Signed-off-by: Brian J. Murrell <brian.murrell@intel.com>